+++ discordiador/include/ops_tripulante.h
@@ -17,14 +17,14 @@ sem_t BLOQUEAR_PLANIFICADOR;
 
 void cerrar_conexiones_tripulante(t_tripulante*);
 void free_t_tripulante(void*);
-t_tripulante* iniciar_tripulante(t_posicion*, uint16_t);
+t_tripulante* iniciar_tripulante(t_posicion*, uint32_t);
 uint8_t solicitar_tarea(t_tripulante*);
-//uint8_t op_expulsar_tripulante(uint16_t);
+//uint8_t op_expulsar_tripulante(uint32_t);
 
 void bloquear_planificacion();
 void reanudar_planificacion();
 
-uint16_t generar_pid();
-uint16_t generar_tid();
+uint32_t generar_pid();
+uint32_t generar_tid();
 
 #endif
